platformaDoKursow/forms/chapter_form.py
from django import forms
from ckeditor.widgets import CKEditorWidget
from ckeditor_uploader.widgets import CKEditorUploadingWidget
from platformaDoKursow.models import Chapter

class ChapterForm(forms.ModelForm):
    content = forms.CharField(widget=CKEditorUploadingWidget())

    class Meta:
        model = Chapter
        fields = ['title', 'content']
