requirements.txt
amqp==5.1.1
asgiref==3.3.4
attrs==22.1.0
billiard==3.6.4.0
boto3==1.24.70
botocore==1.27.70
cachetools==5.2.0
celery==5.2.2
certifi==2022.6.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==6.4.4
cssselect==1.1.0
cssutils==2.6.0
Django==3.2
django-celery-beat==2.2.1
django-cors-headers==3.13.0
django-filter==22.1
django-log-viewer==1.1.6
django-reset-migrations==0.4.0
django-tables2==2.4.0
django-timezone-field==4.2.3
djangorestframework==3.13.1
fsspec==2022.8.2
grpcio==1.47.0
grpcio-status==1.47.0
idna==3.3
iniconfig==1.1.1
jmespath==1.0.1
joblib==1.1.0
kombu==5.2.4
lxml==4.9.1
mysqlclient==2.1.1
nltk==3.7
numpy==1.23.3
oauthlib==3.2.1
packaging==21.3
pandas==1.4.4
pluggy==1.0.0
premailer==3.10.0
prompt-toolkit==3.0.31
proto-plus==1.19.6
protobuf==3.20.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMySQL==1.0.2
pyparsing==3.0.9
pytest==7.1.3
pytest-django==4.5.2
pytest-dotenv==0.5.2
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytest-factoryboy==2.5.0
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
regex==2022.9.11
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
s3fs==0.4.2
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==1.4.41
sqlparse==0.4.1
tomli==2.0.1
tqdm==4.64.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
yagmail==0.15.293
django-bootstrap-v5
django-ckeditor
